moved question to model folder
