+++ src/models/question.ts
@@ -1,0 +1,8 @@
+interface question {
+  body: string,
+  name: string,
+  email: string,
+  "product_id": number,
+}
+
+export default question;
